+++ lib/utils/logical_reformatter.hpp
@@ -16,9 +16,9 @@ public:
 private:
   std::vector<std::vector<int>>
   getProblemMatrix(const core::InputRows input_rows);
-  std::vector<double> getLowerBounds(const core::InputRows input_rows);
-  std::vector<double> getUpperBounds(const core::InputRows input_rows,
-                                     const std::vector<double> lower_bounds);
+  std::vector<int> getLowerBounds(const core::InputRows input_rows);
+  std::vector<int> getUpperBounds(const core::InputRows input_rows,
+                                  const std::vector<int> lower_bounds);
 };
 
 } // namespace utils
